requirements.txt
django
langchain
langchain-openai
faiss-cpu
pypdf
python-dotenv